// src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const Message = __t.object("Message", {
  id: __t.u64(),
  roomId: __t.u64(),
  sender: __t.identity(),
  sent: __t.timestamp(),
  text: __t.string(),
  editedAt: __t.option(__t.timestamp()),
  replyToId: __t.option(__t.u64()),
  imageUrl: __t.option(__t.string()),
});
export type Message = __Infer<typeof Message>;

export const Reaction = __t.object("Reaction", {
  id: __t.u64(),
  messageId: __t.u64(),
  sender: __t.identity(),
  emoji: __t.string(),
});
export type Reaction = __Infer<typeof Reaction>;

export const Room = __t.object("Room", {
  id: __t.u64(),
  name: __t.string(),
  isDm: __t.bool(),
  createdBy: __t.identity(),
  createdAt: __t.timestamp(),
});
export type Room = __Infer<typeof Room>;

export const RoomMember = __t.object("RoomMember", {
  id: __t.u64(),
  roomId: __t.u64(),
  userId: __t.identity(),
  joinedAt: __t.timestamp(),
  lastReadMessageId: __t.u64(),
});
export type RoomMember = __Infer<typeof RoomMember>;

export const User = __t.object("User", {
  identity: __t.identity(),
  name: __t.option(__t.string()),
  online: __t.bool(),
  typing: __t.bool(),
  bio: __t.option(__t.string()),
  avatarUrl: __t.option(__t.string()),
});
export type User = __Infer<typeof User>;